Passwords, comments, and email
User
Form to change the password via AJAX.

Comment
Form for the registration of answers
Query and display all responses in a module

Email sending
Using PHPMailler it is possible to send email to all users involved in the resolution of the module

CKEDITOR Code Snippet
Integrated code editor, so you can insert and format source code snippets directly into the editor.
On the modules page, you can view the formatted codes.